alterando anúncios para dados globais
